requirements.txt

python-telegram-bot
google-genai
pillow

